pyproject.toml
[project]
name = "geminiplayground"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "alive-progress>=3.2.0",
    "arrow>=1.3.0",
    "diskcache>=5.6.3",
    "fastapi[standard]>=0.115.11",
    "gitpython>=3.1.44",
    "google-genai>=1.7.0",
    "greenlet>=3.1.1",
    "langchain>=0.3.21",
    "langchain-core>=0.3.47",
    "langchain-google-genai>=2.1.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "pygithub>=2.6.1",
    "pymupdf>=1.25.4",
    "rich>=13.9.4",
    "tenacity>=9.0.0",
    "tqdm>=4.67.1",
    "validators>=0.34.0",
    "websockets>=15.0.1",
    "yaspin>=3.1.0",
]

[project.optional-dependencies]
demos = [
    "langchain-weaviate>=0.0.4",
]


[tool.pdm.scripts]
geminiplayground = "geminiplayground.cli:app"

[project.scripts]
geminiplayground = "geminiplayground.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
    "tests/",
    "docs/",
    "examples/",
    "assets/",
    "keys/",
    ".git/",
    ".github/",
    "scripts/",
    "configs/",
    "README.md",
    ".env",
    "config.json",
    "*/**/.sh",
    "Makefile",
    "cloudbuild.yaml",
    "*/**/Dockerfile",
    "Dockerfile",
    "requirements.txt",
    "docker/",
    "*.sh",
    "*.yml",
    "analysis/",
    "data/",
    "images/",
    "ui/",
    "web.ps1",
]
